requirements.txt
flask
gunicorn
requests
google-auth
google-auth-oauthlib
google-auth-httplib2
google-api-python-client
gspread
secure-smtplib
datetime
oauth2client
flask_caching
